Support features with partially numeric names
Some of Modernizr's available features have digits in their names, but gulp-modernizr-build previously incorrectly assumed that all feature names contain only the characters a-z, /, and -. This commit simply adds 0-9 to the regex.

For example, my project uses the 'es5' feature, and I was very confused when gulp-modernizr-build crashed out from not finding it. ;)